requirements.txt
# HTTP 请求
requests>=2.31.0

# 邮件发送
resend>=0.8.0

# 模板引擎
jinja2>=3.1.3

# 日期处理
python-dateutil>=2.8.2

# 环境变量
python-dotenv>=1.0.0
